Restructure and Rev3
I restructured the Github directory to follow the Systems Engineering
format.  I also updated the VIPB to contain the examples.  It should be
noted that in order to build the examples into the package, I needed to
remove the CPU monitoring.  The CPU monitoring is not a part of the base
vi.lib, and I'm not sure how to include it.  We may look into this
later.